added SecurityUtil, added methods to services, small refactor

+++ src/main/java/ru/nfm/calendar/service/JwtRefreshTokenService.java
@@ -4,7 +4,7 @@ import ru.nfm.calendar.model.RefreshToken;
 
 public interface JwtRefreshTokenService {
 
-    RefreshToken createRefreshToken(Long userId);
+    RefreshToken createRefreshToken(int userId);
 
     RefreshToken verifyExpiration(RefreshToken refreshToken);
 }
