+++ src/main/java/ru/nfm/calendar/service/impl/JwtRefreshTokenServiceImpl.java
@@ -26,7 +26,7 @@ public class JwtRefreshTokenServiceImpl implements JwtRefreshTokenService {
 
     @Override
     @Transactional
-    public RefreshToken createRefreshToken(Long userId) {
+    public RefreshToken createRefreshToken(int userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UserNotFoundException(userId, "User not found"));
         RefreshToken refreshToken = RefreshToken.builder()
